合并至20240520 Lavac Hello World！
A=增加, B=修复, C=更改, D=删除

更新Readme
A:
	Minecraft 1.19.2 的服务器协议和3D旋转验证码
	Lavac:
		新增TypeResolver单独处理类型解析
		实现部分API
		加入了一键式的LavaCompiler，编译字符串形式的lambda
	FRP:
		使用Blake3替代不安全的SHA1哈希
		加入了nogui和autorun参数
	为FieldAccessor#getObject增加了基本类型自动包装
	基于ASM的基本类型数组Hasher
	NovelFrame支持循环替换
	DynByteBuf加入hex()和base64()
	基于ASM的事件总线，可继承，可取消注册，并且有错误提示
	加入Asn.1和Der编码的支持
	加入Jar完整性检查
	加入DateParser，方便的解析日期
	加入更高性能的TokenizerX实现便于未来替换Tokenizer
	加入了完美哈希表的实现
	加入了Autoload注解来自动加载类、Nixim或转换器
	加入了两种支持增量修改的分卷7z压缩格式
	加入了QOI图片格式的编码和解码
	QZArchiver现已支持尾部追加模式，在符合条件的情况下，可以减少硬盘写入并提高速度

B:
	定时任务可能意外的取消
	AnsiString#isBgColorRGB错误判断
	完成重构到一半的PacketBuffer
	Pipe本地连接CPU占用率高
	CCYaml和YAMLParser代码不同步
	Feedback CFB未遵循规范
	TextWriter性能很差
	BottomLine无法显示
	随机崩溃（懒常量解析）
	fastZip文件名处理错误（URIUtil.unescapeFileName）
	Mapper#loadCache未正确调用packup
	分卷压缩文件识别错误

C:
	ModuleKiller可以直接运行
	AnnotationRepo仅存储AccessData以降低内存占用
	Serializer: 不开启对象池的序列化会继续使用传统模式处理数组和列表
	FRP: 加入哈希白名单和自动登录功能
	重构日志记录器上下文
	HTTP: 重构定义式路由器，现在支持Vue-Router的路由语法
	修改插件加载方式，现在能在卸载时取消定时任务、注册的指令和HTTP路由
	重构AnnotationRepo和DeclFilter更易于使用
	整理了roj.collect中大部分的类
	将MutableInt合入CInteger
	将Nixim和ClassDefiner的调试代码移动至ClassDumper
	合并几个CLI工具到Archive4J
	CLI Console: Completion支持向前替换
	让Argument<File>更易于使用（加入了完整的tab补全支持）
	QZArchive的内存占用现已合并计算，避免构造特殊文件可能造成的意外内存溢出
	高精度计时器现在与keepalive daemon合并

D:
	删除了过时的Vinary编码
	删除了用不上的数学类

+++ java/roj/math/Vec4d.java
@@ -21,7 +21,7 @@ public class Vec4d extends Vector {
 	public Vec4d(Vec2d xy, double z, double s) {
 		this(xy.x, xy.y, z, s);
 	}
-	public Vec4d(Vec3f xyz, double s) {
+	public Vec4d(Vec3d xyz, double s) {
 		this(xyz.x, xyz.y, xyz.z, s);
 	}
 	public Vec4d(Vec4d xyzw) {
@@ -35,7 +35,7 @@ public class Vec4d extends Vector {
 		this.s = w;
 		return this;
 	}
-	public Vec4d set(Vec3f xyz, double w) {
+	public Vec4d set(Vec3d xyz, double w) {
 		this.x = xyz.x;
 		this.y = xyz.y;
 		this.z = xyz.z;
